test(tabs): migrate basic
